+++ CHANGES.md
@@ -4,6 +4,8 @@ Unreleased
 * atdgen: Breaking change, migrate from Bucklescript to Melange (#375)
 * atdd: Workaround d compiler bug regarding declaration order when using aliases (#393)
         Algebraic data types (SumType) now uses `alias this` syntax.
+* atdd: Support recursive records, recursive variants through pointers.
+        Support using enum for variant with no data.
 
 2.15.0 (2023-10-26)
 -------------------